requirements.txt
fastapi[standard]
uvicorn[standard]
authlib
httpx
python-multipart
sqlalchemy
alembic
asyncpg
pydantic
pydantic_settings
passlib[bcrypt]
python-jose[cryptography]
celery[redis]
redis
starlette
python-dotenv
itsdangerous
requests